Updated to 1.4.2
git-svn-id: https://plugins.svn.wordpress.org/pinterest-pin-it-button/trunk@633342 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -5,16 +5,26 @@ if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-//Remove option records from options table
-delete_option( 'pib_options' );
-delete_option( 'pib_category_fields_option' );
+//Delete plugin options only if value checked
 
-//Remove custom post meta fields
-$posts = get_posts( array( 'numberposts' => -1 ) );
+global $pib_options;
 
-foreach( $posts as $post ) {
-    delete_post_meta( $post->ID, 'pib_sharing_disabled' );
-    delete_post_meta( $post->ID, 'pib_url_of_webpage' );
-    delete_post_meta( $post->ID, 'pib_url_of_img' );
-    delete_post_meta( $post->ID, 'pib_description' );
+//Need to retrieve options here
+$pib_options = get_option( 'pib_options' );
+
+if ( !(bool)$pib_options['uninstall_save_settings'] ) {
+
+    //Remove option records from options table
+    delete_option( 'pib_options' );
+    delete_option( 'pib_category_fields_option' );
+
+    //Remove custom post meta fields
+    $posts = get_posts( array( 'numberposts' => -1 ) );
+
+    foreach( $posts as $post ) {
+        delete_post_meta( $post->ID, 'pib_sharing_disabled' );
+        delete_post_meta( $post->ID, 'pib_url_of_webpage' );
+        delete_post_meta( $post->ID, 'pib_url_of_img' );
+        delete_post_meta( $post->ID, 'pib_description' );
+    }
 }
